Refactor: Restructure project for improved maintainability
This commit reorganizes the project structure as follows:

-   **Configuration:**
    -   `Constants.h` moved to a new top-level `config/` directory.

-   **Source Code (`src/`):**
    -   Core game logic (`Game.cpp`) moved to `src/core/`.
    -   Entity implementations (`Player.cpp`, `Wall.cpp`) moved to `src/entities/`.

-   **Header Files (`include/`):**
    -   `Game.h` moved to `include/core/`.
    -   `Player.h` and `Wall.h` moved to `include/entities/`.
    -   (Note: Headers were originally found in `src/` and moved to their respective new locations under `include/`)

-   **Build System (`CMakeLists.txt`):**
    -   Updated to reflect new file paths for sources.
    -   `include_directories` now includes `config` and `include` (allowing for includes like `core/Game.h` or `entities/Player.h`).
    -   Source file list in `CMakeLists.txt` grouped for better readability.

All `#include` directives in source and header files have been updated to match the new structure.

**Note:** I was unable to fully verify the build and run steps due to persistent environment errors that prevented CMake from executing successfully. The changes are based on successful file operations and path updates.
